MovieDetail shows proper star if movie is marked as favorite

+++ app/src/main/java/com/example/android/popularmovies/MovieInfo.java
@@ -7,7 +7,7 @@ import java.io.Serializable;
  */
 
 public class MovieInfo implements Serializable {
-    public int tmdbId;
+    public long tmdbId;
     public String overview;
     public String posterPath;
     public String releaseDate;
